@@ -1,4 +1,4 @@
 package com.philipphecht;
 
-public class FileProviderClass extends android.support.v4.content.FileProvider {
+public class FileProviderClass extends androidx.core.content.FileProvider {
 }
